move object to new namespace.
