+++ osu.Game.Rulesets.Karaoke.Tests/UI/TestSceneControlLayer.cs
@@ -7,7 +7,7 @@ using osu.Framework.Graphics;
 using osu.Framework.Graphics.Containers;
 using osu.Game.Rulesets.Karaoke.Configuration;
 using osu.Game.Tests.Visual;
-using static osu.Game.Rulesets.Karaoke.UI.Overlays.SettingHUDOverlay;
+using static osu.Game.Rulesets.Karaoke.UI.HUD.SettingHUDOverlay;
 
 namespace osu.Game.Rulesets.Karaoke.Tests.UI
 {
