+++ osu.Game.Rulesets.Karaoke/Mods/IApplicableToSettingHUDOverlay.cs
@@ -1,7 +1,7 @@
 ﻿// Copyright (c) andy840119 <andy840119@gmail.com>. Licensed under the GPL Licence.
 // See the LICENCE file in the repository root for full licence text.
 
-using osu.Game.Rulesets.Karaoke.UI.Overlays;
+using osu.Game.Rulesets.Karaoke.UI.HUD;
 using osu.Game.Rulesets.Mods;
 
 namespace osu.Game.Rulesets.Karaoke.Mods
