+++ osu.Game.Rulesets.Karaoke/Mods/KaraokeModPractice.cs
@@ -10,9 +10,8 @@ using osu.Game.Rulesets.Karaoke.Graphics.Sprites;
 using osu.Game.Rulesets.Karaoke.Objects;
 using osu.Game.Rulesets.Karaoke.Replays;
 using osu.Game.Rulesets.Karaoke.UI;
-using osu.Game.Rulesets.Karaoke.UI.Overlays;
-using osu.Game.Rulesets.Karaoke.UI.Overlays.Settings;
-using osu.Game.Rulesets.Karaoke.UI.Overlays.Settings.PlayerSettings;
+using osu.Game.Rulesets.Karaoke.UI.HUD;
+using osu.Game.Rulesets.Karaoke.UI.PlayerSettings;
 using osu.Game.Rulesets.Mods;
 using osu.Game.Rulesets.UI;
 using osu.Game.Scoring;
