+++ osu.Game.Rulesets.Karaoke/UI/HUD/PopupOverlay.cs
@@ -9,7 +9,7 @@ using osu.Game.Graphics;
 using osu.Game.Overlays;
 using osuTK.Graphics;
 
-namespace osu.Game.Rulesets.Karaoke.UI.Overlays.Settings
+namespace osu.Game.Rulesets.Karaoke.UI.HUD
 {
     public abstract class PopupOverlay : WaveOverlayContainer
     {
