+++ osu.Game.Rulesets.Karaoke/UI/HUD/SettingButton.cs
@@ -5,7 +5,7 @@ using osu.Framework.Graphics.Cursor;
 using osu.Framework.Localisation;
 using osu.Game.Graphics.UserInterface;
 
-namespace osu.Game.Rulesets.Karaoke.UI.Overlays.Settings
+namespace osu.Game.Rulesets.Karaoke.UI.HUD
 {
     public class SettingButton : TriangleButton, IHasTooltip
     {
