+++ osu.Game.Rulesets.Karaoke/UI/PlayerSettings/PitchSettings.cs
@@ -7,7 +7,7 @@ using osu.Game.Graphics.Sprites;
 using osu.Game.Rulesets.Karaoke.Configuration;
 using osu.Game.Screens.Play.PlayerSettings;
 
-namespace osu.Game.Rulesets.Karaoke.UI.Overlays.Settings.PlayerSettings
+namespace osu.Game.Rulesets.Karaoke.UI.PlayerSettings
 {
     public class PitchSettings : PlayerSettingsGroup
     {
