+++ osu.Game.Rulesets.Karaoke/UI/PlayerSettings/RubyRomajiSettings.cs
@@ -6,7 +6,7 @@ using osu.Framework.Graphics;
 using osu.Game.Rulesets.Karaoke.Configuration;
 using osu.Game.Screens.Play.PlayerSettings;
 
-namespace osu.Game.Rulesets.Karaoke.UI.Overlays.Settings.PlayerSettings
+namespace osu.Game.Rulesets.Karaoke.UI.PlayerSettings
 {
     public class RubyRomajiSettings : PlayerSettingsGroup
     {
