+++ osu.Game.Rulesets.Karaoke/UI/PlayerSettings/TranslateSettings.cs
@@ -9,7 +9,7 @@ using osu.Game.Graphics.UserInterface;
 using osu.Game.Rulesets.Karaoke.Configuration;
 using osu.Game.Screens.Play.PlayerSettings;
 
-namespace osu.Game.Rulesets.Karaoke.UI.Overlays.Settings.PlayerSettings
+namespace osu.Game.Rulesets.Karaoke.UI.PlayerSettings
 {
     public class TranslateSettings : PlayerSettingsGroup
     {
